Struct gxi::proto::ThemeSettings

pub struct ThemeSettings {
    pub foreground: Option<Color>,
    pub background: Option<Color>,
    pub caret: Option<Color>,
    pub line_highlight: Option<Color>,
    pub misspelling: Option<Color>,
    pub minimap_border: Option<Color>,
    pub accent: Option<Color>,
    pub popup_css: Option<String>,
    pub phantom_css: Option<String>,
    pub bracket_contents_foreground: Option<Color>,
    pub bracket_contents_options: Option<UnderlineOption>,
    pub brackets_foreground: Option<Color>,
    pub brackets_background: Option<Color>,
    pub brackets_options: Option<UnderlineOption>,
    pub tags_foreground: Option<Color>,
    pub tags_options: Option<UnderlineOption>,
    pub highlight: Option<Color>,
    pub find_highlight: Option<Color>,
    pub find_highlight_foreground: Option<Color>,
    pub gutter: Option<Color>,
    pub gutter_foreground: Option<Color>,
    pub selection: Option<Color>,
    pub selection_foreground: Option<Color>,
    pub selection_background: Option<Color>,
    pub selection_border: Option<Color>,
    pub inactive_selection: Option<Color>,
    pub inactive_selection_foreground: Option<Color>,
    pub guide: Option<Color>,
    pub active_guide: Option<Color>,
    pub stack_guide: Option<Color>,
    pub highlight_foreground: Option<Color>,
    pub shadow: Option<Color>,
}

Various properties meant to be used to style a text editor. Basically all the styles that aren't directly applied to text like selection colour. Use this to make your editor UI match the highlighted text.

Fields
foreground: Option<Color>

The default color for text.

background: Option<Color>

The default backgound color of the view.

caret: Option<Color>

Color of the caret.

line_highlight: Option<Color>

Color of the line the caret is in. Only used when the higlight_line setting is set to true.

misspelling: Option<Color>

The color to use for the squiggly underline drawn under misspelled words.

minimap_border: Option<Color>

The color of the border drawn around the viewport area of the minimap. Only used when the draw_minimap_border setting is enabled.

accent: Option<Color>

A color made available for use by the theme.

popup_css: Option<String>

CSS passed to popups.

phantom_css: Option<String>

CSS passed to phantoms.

bracket_contents_foreground: Option<Color>

Color of bracketed sections of text when the caret is in a bracketed section. Only applied when the match_brackets setting is set to true.

bracket_contents_options: Option<UnderlineOption>

Controls certain options when the caret is in a bracket section. Only applied when the match_brackets setting is set to true.

brackets_foreground: Option<Color>

Foreground color of the brackets when the caret is next to a bracket. Only applied when the match_brackets setting is set to true.

brackets_background: Option<Color>

Background color of the brackets when the caret is next to a bracket. Only applied when the match_brackets setting is set to true.

brackets_options: Option<UnderlineOption>

Controls certain options when the caret is next to a bracket. Only applied when the match_brackets setting is set to true.

tags_foreground: Option<Color>

Color of tags when the caret is next to a tag. Only used when the match_tags setting is set to true.

tags_options: Option<UnderlineOption>

Controls certain options when the caret is next to a tag. Only applied when the match_tags setting is set to true.

highlight: Option<Color>

The border color for "other" matches.

find_highlight: Option<Color>

Background color of regions matching the current search.

find_highlight_foreground: Option<Color>

Text color of regions matching the current search.

gutter: Option<Color>

Background color of the gutter.

gutter_foreground: Option<Color>

Foreground color of the gutter.

selection: Option<Color>

The background color of selected text.

selection_foreground: Option<Color>

A color that will override the scope-based text color of the selection.

selection_background: Option<Color>

Deprecated!

This property is not part of the recognized tmTheme format. It may be removed in a future release.

selection_border: Option<Color>

Color of the selection regions border.

inactive_selection: Option<Color>

The background color of a selection in a view that is not currently focused.

inactive_selection_foreground: Option<Color>

A color that will override the scope-based text color of the selection in a view that is not currently focused.

guide: Option<Color>

Color of the guides displayed to indicate nesting levels.

active_guide: Option<Color>

Color of the guide lined up with the caret. Only applied if the indent_guide_options setting is set to draw_active.

stack_guide: Option<Color>

Color of the current guide’s parent guide level. Only used if the indent_guide_options setting is set to draw_active.

highlight_foreground: Option<Color>

Foreground color for regions added via sublime.add_regions() with the sublime.DRAW_OUTLINED flag added.

Deprecated! This setting does not exist in any available documentation. Use is discouraged, and it may be removed in a future release.

shadow: Option<Color>

The color of the shadow used when a text area can be horizontally scrolled.