Struct gxi::edit_view::EditView

pub struct EditView {
    core: Rc<RefCell<Core>>,
    main_state: Rc<RefCell<MainState>>,
    pub view_id: String,
    pub file_name: Option<String>,
    pub pristine: bool,
    pub da: DrawingArea,
    pub root_widget: Box,
    pub tab_widget: Box,
    search_bar: SearchBar,
    search_entry: SearchEntry,
    find_status_label: Label,
    pub label: Label,
    pub close_button: Button,
    hscrollbar: Scrollbar,
    vscrollbar: Scrollbar,
    line_cache: LineCache,
    replace: EVReplace,
    font: EVFont,
}

The EditView is the part of gxi that does the actual editing. This is where you edit documents.

Fields
core: Rc<RefCell<Core>> main_state: Rc<RefCell<MainState>> view_id: String file_name: Option<String> pristine: bool da: DrawingArea root_widget: Box tab_widget: Box search_bar: SearchBar search_entry: SearchEntry find_status_label: Label label: Label close_button: Button hscrollbar: Scrollbar vscrollbar: Scrollbar line_cache: LineCache replace: EVReplace font: EVFont
Methods

impl EditView

pub fn new(
main_state: &Rc<RefCell<MainState>>,
core: &Rc<RefCell<Core>>,
file_name: Option<String>,
view_id: &str
) -> Rc<RefCell<EditView>>

Initialises a new EditView. Sets up scrollbars, the actual editing area, the fonts, the syntax lang and connects all events which might happen during usage (e.g. scrolling)

impl EditView

pub fn set_file(&mut self, file_name: &str)

Set the name of the file the EditView is currently editing and calls update_title

fn update_title(&self)

Update the title of the EditView to the currently set file_name

pub fn config_changed(&mut self, changes: &Value)

If xi-editor sends us a config_changed msg we process it here, e.g. setting the font face/size xi-editor tells us. Most configs don't need special handling by us though.

pub fn update(&mut self, params: &Value)

If xi-editor sends us a update msg we process it here, setting the scrollbars upper limit accordingly, checking if the EditView is pristine (has unsaved changes) and queue a new draw of the EditView.

fn change_scrollbar_visibility(&self)

pub fn da_px_to_cell(
&self,
main_state: &MainState,
x: f64,
y: f64
) -> (u64, u64)

Maps x|y pixel coordinates to the line num and col. This can be used e.g. for determining the firt and last time, but setting the y coordinate to 0 and the last pixel.

fn da_size_allocate(&mut self, da_width: i32, da_height: i32)

Allocate the space our DrawingArea needs.

fn vscrollbar_change_value(&mut self, value: f64) -> Inhibit

Upon changing the vertical scrollbar we have to call update_visible_scroll_region

fn update_visible_scroll_region(&self)

This updates the part of the document that's visible to the user, e.g. when scrolling. This requests the required lines from xi-editor to add them to the line cache and then adjusts the scrolling to the visible region.

fn get_text_size(&self) -> (f64, f64)

Returns the width&height of the entire document

pub fn handle_draw(&mut self, cr: &Context) -> Inhibit

Handles the drawing of the EditView. This is called when we get a update from xi-editor or if gtk requests us to draw the EditView. This draws the background, all lines and the cursor.

fn create_layout_for_linecount(
&self,
pango_ctx: &Context,
_main_state: &MainState,
n: u64,
padding: usize
) -> Layout

Creates a pango layout for a particular linecount (the count on the left) in the linecache

pub fn line_width(&self, line_string: &str) -> f64

Checks how wide a line is

fn create_layout_for_line(
&self,
pango_ctx: &Context,
main_state: &MainState,
line: &Line
) -> Layout

Creates a pango layout for a particular line in the linecache

pub fn scroll_to(&mut self, line: u64, col: u64)

Scrolls vertically to the line specified and horizontally to the column specified.

pub fn handle_button_press(&self, eb: &EventButton) -> Inhibit

Handles button presses such as Shift, Ctrl etc. and primary pasting (i.e. via Ctrl+V, not via middle mouse click).

pub fn handle_drag(&mut self, em: &EventMotion) -> Inhibit

Handle selecting line(s) by dragging the mouse across them while having the left mouse button clicked.

pub fn handle_scroll(&mut self, es: &EventScroll) -> Inhibit

Handles scroll events, i.e. the user dragging the scrollbar, scrolling via a mouse wheel or via a touchpad/drawing tablet (which use SmoothScrolling, which may scroll vertically and horizontally at the same time).

fn handle_key_press_event(&mut self, ek: &EventKey) -> Inhibit

Handles all (special) key press events, e.g. copy, pasting, PgUp/Down etc.

fn do_cut(&self, view_id: &str)

Copies text to the clipboard

fn do_copy(&self, view_id: &str)

Copies text to the clipboard

fn do_paste(&self, view_id: &str)

Pastes text from the clipboard into the EditView

fn do_paste_primary(&self, view_id: &str, line: u64, col: u64)

fn do_resize(&self, view_id: &str, width: i32, height: i32)

Resize the EditView

pub fn start_search(&self)

Opens the find dialog (Ctrl+F)

pub fn start_replace(&self)

Opens the replace dialog (Ctrl+R)

pub fn stop_search(&self)

Closes the find/replace dialog

pub fn find_status(&self, queries: &Value)

Displays how many matches have been found in the find/replace dialog.

pub fn replace_status(&self, status: &Value)

Displays what chars will be replaced in the replace dialog

pub fn find_next(&self)

Go to the next match in the find/replace dialog

pub fn find_prev(&self)

Go the to previous match in the find/replace dialog

pub fn search_changed(&self, s: Option<String>)

Tells xi-editor that we're searching for a different string (or none) now

pub fn replace(&self)

Replace one match with the replacement string

pub fn replace_all(&self)

Replace all matches with the replacement string