Struct gxi::rpc::Core

pub struct Core {
    pub state: Arc<Mutex<CoreState>>,
}

Fields
state: Arc<Mutex<CoreState>>
Methods

impl Core

pub fn new(
xi_peer: XiPeer,
xi_rx: Receiver<Value>,
err_tx: Sender<ErrMsg>,
shared_queue: SharedQueue
) -> Self

Sets up a new RPC connection, also starting a thread to receive responses.

The handler is invoked for incoming RPC notifications. Note that it must be Send because it is called from a dedicated thread.

pub fn send_notification(&self, method: &str, params: &Value)

pub fn send_result(&self, id: u64, result: &Value)

pub fn send_request<F>(&self, method: &str, params: &Value, callback: F) where
F: FnOnce(&Value) + Send + 'static,

Calls the callback with the result (from a different thread).

pub fn save(&self, view_id: &str, file_path: &str)

pub fn close_view(&self, view_id: &str)

fn send_edit_cmd(&self, view_id: &str, method: &str, params: &Value)

pub fn client_started(&self, config_dir: &str, client_extras_dir: &str)

pub fn set_theme(&self, theme_name: &str)

pub fn insert(&self, view_id: &str, chars: &str)

Inserts the chars string at the current cursor location.

pub fn resize(&self, view_id: &str, width: i32, height: i32)

pub fn delete_forward(&self, view_id: &str)

pub fn delete_backward(&self, view_id: &str)

pub fn insert_newline(&self, view_id: &str)

pub fn insert_tab(&self, view_id: &str)

pub fn move_up(&self, view_id: &str)

pub fn move_down(&self, view_id: &str)

pub fn move_left(&self, view_id: &str)

pub fn move_right(&self, view_id: &str)

pub fn move_up_and_modify_selection(&self, view_id: &str)

pub fn move_down_and_modify_selection(&self, view_id: &str)

pub fn move_left_and_modify_selection(&self, view_id: &str)

pub fn move_right_and_modify_selection(&self, view_id: &str)

pub fn move_word_left(&self, view_id: &str)

pub fn move_word_right(&self, view_id: &str)

pub fn move_word_left_and_modify_selection(&self, view_id: &str)

pub fn move_word_right_and_modify_selection(&self, view_id: &str)

pub fn move_to_left_end_of_line(&self, view_id: &str)

pub fn move_to_right_end_of_line(&self, view_id: &str)

pub fn move_to_left_end_of_line_and_modify_selection(&self, view_id: &str)

pub fn move_to_right_end_of_line_and_modify_selection(&self, view_id: &str)

pub fn move_to_beginning_of_document(&self, view_id: &str)

pub fn move_to_end_of_document(&self, view_id: &str)

pub fn move_to_beginning_of_document_and_modify_selection(&self, view_id: &str)

pub fn move_to_end_of_document_and_modify_selection(&self, view_id: &str)

pub fn page_up(&self, view_id: &str)

pub fn page_down(&self, view_id: &str)

pub fn page_up_and_modify_selection(&self, view_id: &str)

pub fn page_down_and_modify_selection(&self, view_id: &str)

pub fn select_all(&self, view_id: &str)

pub fn gesture_point_select(&self, view_id: &str, line: u64, col: u64)

moves the cursor to a point (click)

pub fn gesture_toggle_sel(&self, view_id: &str, line: u64, col: u64)

adds or removes a selection at a point (new cursor)

pub fn gesture_range_select(&self, view_id: &str, line: u64, col: u64)

modifies the selection to include a point (shift+click)

pub fn gesture_line_select(&self, view_id: &str, line: u64, col: u64)

sets the selection to a given line (triple click)

pub fn gesture_word_select(&self, view_id: &str, line: u64, col: u64)

sets the selection to a given word (double click)

pub fn gesture_multi_line_select(&self, view_id: &str, line: u64, col: u64)

adds a line to the selection

pub fn gesture_multi_word_select(&self, view_id: &str, line: u64, col: u64)

adds a word to the selection

pub fn scroll(&self, view_id: &str, first: u64, last: u64)

Notifies the back-end of the visible scroll region, defined as the first and last (non-inclusive) formatted lines. The visible scroll region is used to compute movement distance for page up and page down commands, and also controls the size of the fragment sent in the update method.

pub fn drag(&self, view_id: &str, line: u64, col: u64, modifier: u32)

pub fn undo(&self, view_id: &str)

pub fn redo(&self, view_id: &str)

pub fn cut(&mut self, view_id: &str) -> Option<String>

pub fn copy(&mut self, view_id: &str) -> Option<String>

pub fn paste(&self, view_id: &str, chars: &str)

pub fn find(
&self,
view_id: &str,
chars: &str,
case_sensitive: bool,
regex: Option<bool>
)

Searches the document for chars, if present, falling back on the last selection region if chars is None.

If chars is None and there is an active selection, returns the string value used for the search, else returns Null.

pub fn find_next(
&self,
view_id: &str,
wrap_around: Option<bool>,
allow_same: Option<bool>
)

pub fn find_previous(&self, view_id: &str, wrap_around: Option<bool>)

pub fn highlight_find(&self, view_id: &str, visible: bool)

pub fn replace(&self, view_id: &str, chars: &str, preserve_case: bool)

pub fn replace_next(&self, view_id: &str)

pub fn replace_all(&self, view_id: &str)

pub fn set_language(&self, view_id: &str, lang_name: &str)